+++ client/python/tests/test_privileges_command.py
@@ -209,3 +209,74 @@ class TestPrivilegesCommand(CLITestBase):
             ["privileges", "list", "--catalog", "foo", "--catalog-role", "bar"],
         )
         mock_client.list_grants_for_catalog_role.assert_called_with("foo", "bar")
+
+    def test_privilege_catalog_grant(self) -> None:
+        mock_client = self.build_mock_client()
+        self.mock_execute(
+            mock_client,
+            [
+                "privileges",
+                "catalog",
+                "grant",
+                "--catalog",
+                "foo",
+                "--catalog-role",
+                "bar",
+                "TABLE_READ_DATA",
+            ],
+        )
+        call_args = mock_client.add_grant_to_catalog_role.call_args[0]
+        self.assertEqual(call_args[0], "foo")
+        self.assertEqual(call_args[1], "bar")
+        self.assertEqual(call_args[2].grant.privilege.value, "TABLE_READ_DATA")
+
+    def test_privilege_namespace_revoke(self) -> None:
+        mock_client = self.build_mock_client()
+        self.mock_execute(
+            mock_client,
+            [
+                "privileges",
+                "namespace",
+                "revoke",
+                "--namespace",
+                "a.b.c",
+                "--catalog",
+                "foo",
+                "--catalog-role",
+                "bar",
+                "TABLE_READ_DATA",
+            ],
+        )
+        call_args = mock_client.revoke_grant_from_catalog_role.call_args[0]
+        self.assertEqual(call_args[0], "foo")
+        self.assertEqual(call_args[1], "bar")
+        self.assertEqual(call_args[2], False)
+        self.assertEqual(call_args[3].grant.privilege.value, "TABLE_READ_DATA")
+        self.assertEqual(call_args[3].grant.namespace, ["a", "b", "c"])
+
+    def test_privilege_view_revoke(self) -> None:
+        mock_client = self.build_mock_client()
+        self.mock_execute(
+            mock_client,
+            [
+                "privileges",
+                "view",
+                "revoke",
+                "--namespace",
+                "a.b.c",
+                "--catalog",
+                "foo",
+                "--catalog-role",
+                "bar",
+                "--view",
+                "v",
+                "VIEW_FULL_METADATA",
+            ],
+        )
+        call_args = mock_client.revoke_grant_from_catalog_role.call_args[0]
+        self.assertEqual(call_args[0], "foo")
+        self.assertEqual(call_args[1], "bar")
+        self.assertEqual(call_args[2], False)
+        self.assertEqual(call_args[3].grant.privilege.value, "VIEW_FULL_METADATA")
+        self.assertEqual(call_args[3].grant.namespace, ["a", "b", "c"])
+        self.assertEqual(call_args[3].grant.view_name, "v")