+++ persistence/nosql/persistence/cdi/quarkus/src/main/java/org/apache/polaris/persistence/nosql/quarkus/backend/BackendProvider.java
@@ -92,7 +92,7 @@ class BackendProvider {
       var setupSchemaResult = backend.setupSchema().orElse("");
       LOGGER.info("Opened new persistence backend '{}' {}", backend.type(), setupSchemaResult);
 
-      return builder.buildBackend();
+      return backend;
     } catch (Exception e) {
       try {
         backend.close();